removed print statements
***This is the last commit using the easy API in which we assume***
***that the user has access to both galsim and the stack in the***
***same Python***

after this, I will modify the code so that galSimCatalogGenerator.py
writes redshifted, extincted SEDs to a scratch directory.  This will
allow galSimDrawingScript.py to read them in directly and only use
galsim when generating image.  This may be slow because of the I/O
requirements.  It will also require a lot of storage space for large
catalogs, but it is the only way I can think of to get around our
inability to assume that the user will be able to install galsim
with the stack version of Python (a problem for Mac users)