testAllObjects no longer uses unittest.expectedFailure
instead, it uses sys.exc_info() to put a more general filter
on exceptions raised in testAllObjects.py.

now the test should be resilient against failures due to
1) intermittent connectivity
2) lack of db-auth.paf
3) lack of .lsst/

This uncovered an error in GalaxyTileCompoundObj.  We needed
to set doRunTest to false in that class, since it is not an
ordinary CatalogDBObject and should not be tested in the same
way (its __init__ does not accept the kwarg verbose)

messages are printed to the screen regarding how many connections
to fatboy failed and how may succeeded