Bugfix, Aded model to plot, Renamed Example file
Bugfix:  SNObject being read out of catalogs was using wrong units of
ra, dec to calculate MWEBV, and getting non-matching values. This was
resulting in a 1 percent flux difference in cases eyeballed. There was
also a bug where the x0 value was not being set. Both of these were in
the example and not in the catalog method. These bugs are now fixed, and
we demonstrate that fluxes calculated match to high precision. Also we
add the model light curve to the light curve plots, along with pulls.

new file:   SNCatalog_example.ipynb
deleted:    SNCatlog_example.ipynb
Renamed file SNCatlog_example.ipynb to SNCatalog_example.ipynb