added method to de-precess RA, Dec
add method that takes an array of RA, and array of Dec and
an ObservationMetaData.  It calculates the difference between
the ICRS RA, Dec of the ObservationMetaData boresite and the
observed RA, Dec of the boresite.  It then rotates this difference
out of the arrays of input RA, Dec (essentially taking that
transformation out of those coordinates)