moved _dePrecess to PhoSimCatalog classes
In order to avoid users accidentally using _dePrecess without
understanding it, I moved it into a mixin that only the PhoSim
IntanceCatalog classes inherit from.  This should make it clear
that it is a specialized method meant only for use in generating
PhoSim InstanceCatalogs.

This meant moving the getters for raPhoSim and decPhoSim into
those classes, as well, which meant editing a lot of the unit
tests, since _dePrecess, raPhoSim, and decPhoSim are no longer
accessible from AstrometryBase