applyAgn() variability method now caches its results
Previously, whenever you wanted to simulate a variable AGN light
curve, applyAgn() started from the t_0 in the agn variability parameters
and did a damped random walk up to the MJD you asked for.  This meant
that, if you wanted to simulate multiple MJDs, there were time steps that
got simulated repeatedly.  Now, applyAgn stores the results of each random
walk it does and, if it can pick up from where it left off previously, it does.

Note: this required re-defining time step in applyAgn so that it is always
tau/100 (previously, it was defined close to tau/100 but such that the
requested MJD was the exact end of the time grid of time steps). This allows
random walks to be reproducible, even when they are resumed from midway through.