remove unnecessary vectorization from applyAgn
we only ever pass in one MJD at a time to applyAgn, we do not
need it to use numpy vectorization.

This sped up the method by about a factor of two.

Note: it slightly changed the outputs (though, the outputs were
somewhat bogus to begin with).  There was still a disconnect between
dt used to iterate on the damped random walk and dt used to interpolate
the damped random walk and get a final answer.  This commit makes sure
there is only one dt step value.