supernova methods that draw from numpy.random take RandomState arg
this allows us to draw from t0 first, and then decide if we want
to draw from the other parameters, preserving the state of numpy.random
if we do.  I also added a method getSN_rng that returns a RandomState
based on the supernova host's id.  This makes the particular
seeding algorithm used by the supernova catalogs available to users
outside of the context of supernovae catalogs