calculate sne magnitudes with numpy vectorization
Basically, this commit creates a 2D numpy array of SEDs and then
does an end-run around Sed.calcFlux by re-implementing all of that
code in a numpythonic way, so that we don't have to loop over all
of the SEDs in the grid.

Thanks to Rahul Biswas for reminding me that this was a possibility.