Added functions to ObservationMetaDataGenerator for split
The functionality is now split into a function which queries the OpSim
database to obtain the set of desired pointings and a static method
which calculates the ObservationMetaData for each of the pointings.

As a check I have implemented a method called
`getObservationMetaDataNew` which mimics the old
`getObservationMetaData` using split functionality. This will be used in
testing.
	modified:   python/lsst/sims/catUtils/utils/ObservationMetaDataGenerator.py