simplified OpSimGenerator query tests
introduced a method in testObservationMetaDataGenerator.py that maps
between the query commands in getObservationMetaData() and the actual
data stored in an ObservationMetaData.  Used that to shorten the code
used to test that getObservationMetaData() returns values in the correct
ranges.

Note: in order to make these unit tests consistent with our new
API for PhoSimMetaData, an extra column had to be added to the
column mapping in ObservationMetaDataGenerator.py to ensure that
the OpSim angles (all in radians) get mapped to degrees as
PhoSim expects.