"vectorize" applyVariability methods
Pass the full numpy array of objects to the individual methods, rather
than passing objects in one at a time.  This will allow the
applyVariability methods to make use of self.column_by_name().

At this point, not all of the unit tests should pass.  Methods
that do not make use of applyStdPeriodic have not properly been
"vectorized".  I just wanted to commit this code so that I did not
lose my work.

Note: there was a bug in how applyEb() was implemented before.
The period was not properly handled.