Variability mixin constructs its own method registry
This involves several seemingly unrelated changes:

1) remove the testInheritance test from testVariabilityMixins.py.
The method registries are now constructed at instantiation, so this
test would fail.  However, it is redundant with testHybridVariability,
which tests the same things after instantiation, so we are safe to
remove testInheritance.

2) refactor Variability mixin into a base class, which just contains
applyVariability and separate mixins that actually contain the
variability models (e.g. applyRRLyrae, applyAGN, etc.).  This allows
us to pair VariabilityStars (which provides a getter for lsst_u,
lsst_g, etc.) with applyAgn, which is actually an extra galactic
model.  This is what is required, for example, to include AGN
variability in PhoSim catalogs, which will look for delta_lsst_u,
rather than delt_uAgn.