do not multiply by effArea in MLT dwarf flare model
This is because the flux that is used in the MLT flaring
light curves is actually the specific flux density of the source
integrated over the normalized bandpass response (see section 4.1
of LSE-180).  Thus, the flux does not actually depend on the
effective area of the telescope.  It is designed to be agnostic
to everything except the source SED and the shape of the
system throughput.