Fix failures in `testSN.py` following astropy and sncosmo updates
-  Added relative tolerances to tests so that comparisons are not
required to be insanely accurate

- : Fix failure due to wavelengths outside range.

The sncosmo model being used is salt2-extended which has a lower
wavelength range of 1700 Ang. The test required evaluating the
`sncosmo` spectrum for an object at z = 0.96 at a wavelength of 3000
Ang, which corresponds to a rest frame wavelength of 1530.6 Ang. To fix
this problem the self.wave attribute of the test was changed to start
form a lowest wavelength of 3500 Ang.

NOTE: the failure of the test was due to comparing directly with a
`sncosmo` model. `SNObject` models are inensitive to this (and provide
 values of 0 if the spectra are not available, with this being oe of the
 desired changed from `sncosmo`. This would not fly for
cosmology but is better to have rather than end up with failures in
simulation.
	modified:   testSN.py

-  fixing deprecated import for comparison of pandas dataframes
triggering warning
	modified:   testSN.py